requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
